refactor vagrant setup and ubuntu install procedure

+++ install.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 unamestr=`uname`
 lindistro=`cat /etc/issue`
 if [[ $unamestr == Darwin ]]; then
@@ -19,21 +21,36 @@ if [[ $unamestr == Darwin ]]; then
 	#cleanup
 	rm *.dmg
 	#fire-up an Ubuntu VM via vagrant. The
-	cd vagrant
 	echo Installing VM. It may take a while. A cofee may be a good idea now...
 	vagrant up
-	@echo Done. Now run: vagrant ssh
+	echo Done. Now run: vagrant ssh
 elif [[ $unamestr == Linux && $lindistro == Ubuntu* ]]; then
-	LOCALIP=`hostname --ip-address`
-	sudo apt-get -q -y install docker.io
-	curl -L https://github.com/docker/fig/releases/download/1.0.0/fig-`uname -s`-`uname -m` > /tmp/fig; chmod +x /tmp/fig; sudo mv /tmp/fig /usr/local/bin
+	MYDIR=`dirname $0`
+	LOCALIP=${LOCALIP:-$(hostname --ip-address)}
+
+	if [ ! -f /etc/apt/sources.list.d/docker.list ]; then
+		sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 36A1D7869245C8950F966E92D8576A8BA88D21E9
+		echo 'deb http://get.docker.io/ubuntu docker main' > /tmp/docker.list
+		sudo mv /tmp/docker.list /etc/apt/sources.list.d/
+		sudo apt-get update
+	fi
+	sudo apt-get -q -y install lxc-docker
+
+	if [ ! -f /usr/local/bin/fig ]; then
+		curl -L https://github.com/docker/fig/releases/download/1.0.0/fig-`uname -s`-`uname -m` > /tmp/fig; chmod +x /tmp/fig; sudo mv /tmp/fig /usr/local/bin
+	fi
+
+	which docker || { echo "docker command not in path (install failed? check previous output), killing myself"; exit 1; }
+
 	sudo docker stop $(sudo docker ps -a -q)
 	sudo docker stop $(sudo docker ps -a -q)
 	sudo docker rm $(sudo docker ps -a -q)
 	sudo docker rmi -f $(sudo docker images -q)
-	echo Installing VMs. It may take a while. A cofee may be a good idea now...
-        ./build-docker-images.sh
-	HOSTNAME=boot2docker IP=$LOCALIP ./genfig.sh
+	echo Installing docker images. It may take a while. A cofee may be a good idea now...
+	cd $MYDIR
+	sudo ./build-docker-images.sh
+
+	IP=$LOCALIP ./genfig.sh
 	sudo fig up -d
 	echo Done. Now run: sudo docker run -i -t paas
 else
